Change Staff workhours from integer to float #29

+++ src/main/java/es/upm/miw/betca_tpv_spring/documents/Staff.java
@@ -14,13 +14,13 @@ public class Staff {
     private String year;
     private String month;
     private String day;
-    private Integer workHours;
+    private Float workHours;
     private LocalDateTime lastLoginTime;
 
     public Staff() {
     }
 
-    public Staff(String mobile, String year, String month, String day, Integer workHours, LocalDateTime lastLoginTime) {
+    public Staff(String mobile, String year, String month, String day, Float workHours, LocalDateTime lastLoginTime) {
         this.mobile = mobile;
         this.year = year;
         this.month = month;
@@ -79,11 +79,11 @@ public class Staff {
         this.day = day;
     }
 
-    public Integer getWorkHours() {
+    public Float getWorkHours() {
         return workHours;
     }
 
-    public void setWorkHours(Integer workHours) {
+    public void setWorkHours(Float workHours) {
         this.workHours = workHours;
     }
